+++ CHANGELOG.md
@@ -6,6 +6,7 @@ CHANGELOG
 
 * [BC BREAK] Remove `Agent` and `MultiAgent` suffixes from injection aliases.
 * Add `http_client` option to VertexAI platform configuration
+* Add `bedrock` support for platform configuration
 
 0.1
 ---